requirements.txt
python-telegram-bot==13.7
opencv-python-headless
python-dotenv